requirements.txt
streamlit==1.32.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.9.0


langchain>=0.3.6
langchain-google-genai>=2.0.4
langchain-community>=0.3.4
